captcha/v2/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
ddddocr>=1.4.0
Pillow>=10.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
pyyaml>=6.0
openai>=1.0.0
